+++ world/world.gd
@@ -74,6 +74,7 @@ var WORLD_UI_SCENES: Array[PackedScene] = [
 	#preload("res://ui/debug/stream_debug_panel.tscn"),
 	preload("res://ui/debug/perception_debug_panel.tscn"),
 	preload("res://ui/hud/aim_reticle/aim_reticle.tscn"),
+	preload("res://ui/hud/player_hud/player_hud.tscn"),
 ]
 
 const UI_CANVAS_LAYER_INDEX: int = 40